requirements.txt
# Core dependencies
GitPython>=3.1.40
networkx>=3.2.1
PyYAML>=6.0.1
typer>=0.9.0
rich>=13.7.0
openai>=1.0.0
tiktoken>=0.5.0
jsonschema>=4.17.0
tenacity>=8.2.0
python-dotenv>=1.0.1

# Code parsing and analysis
libcst>=1.1.0
astroid>=3.0.1

# LLM Integration
openai>=1.3.5
tiktoken>=0.5.1

# Visualization and reporting
plotly>=5.18.0
jinja2>=3.1.2
graphviz>=0.20.1

# Testing and development
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0

# Type stubs for better type checking
types-PyYAML>=6.0.12.12
types-requests>=2.31.0.10

# Added from the code block
aiohttp==3.9.3
gitpython==3.1.42
jinja2==3.1.3
openai==1.13.3
pytest==8.1.1
pytest-asyncio==0.23.5
weasyprint==61.1
